Refactor: getProviderName 함수 공통 유틸리티로 분리
